+++ src/omv/img/imlib.h
@@ -14,6 +14,7 @@
 #include <arm_math.h>
 #include <ff.h>
 #include "fb_alloc.h"
+#include "umm_malloc.h"
 #include "xalloc.h"
 #include "array.h"
 #include "fmath.h"
